Generate AmigaOS version string in correct format, taking: netsurf_version_major from desktop/version.c as version SVN revision from local repository as revision
This should ensure that the internal version number will increase correctly
(version and revision are treated as seperate values > 0), whereas NetSurf's public
release version (currently coming up to 2.0) is not guaranteed to be in legal AmigaOS
format or increase validly under AmigaOS rules.

The internal version string has been added for future AmiUpdate support as well as
making it easier to see and compare versions of interim builds.

amiga/version.c can safely be removed from the source list when cross-compiling (where
ARexx is not available), or manually created, as it only defines this internal
version string and does not contain any code.

amiga/version.c is automatically removed after linking to ensure it always reflects
the current SVN revision.



git-svn-id: svn://svn.netsurf-browser.org/trunk/netsurf@5510 d004b94f-0d10-0410-95b5-d60e4a4a7a9a