Basic theme support.
NetSurf will use the theme pointed to by the Options file "theme:".  Default is
Resources/Themes/Default

This directory must contain a file called Theme which indicates which files should be
used for which images.  The number of frames in the Throbber must also be specified
here.  See Resources/Themes/Default/Theme for an example.


git-svn-id: svn://svn.netsurf-browser.org/trunk/netsurf@5592 d004b94f-0d10-0410-95b5-d60e4a4a7a9a